refactor: Extract useTemporalCombinedDimension
